bot-twitter/requirements.txt
git+https://github.com/bear/python-twitter
dnspython==1.16.0


python-twitter==3.5
